Some refactoring and a stub of the "Snapshot Events" logic
