+++ source-postgres/testdata/state.json
@@ -1,0 +1,1 @@
+{"SlotName":"estuary_flow_slot","ResumeLSN":23043728}
